requirements.txt
aiogram==2.2
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
arrow==1.3.0
asgiref==3.8.1
attrs==24.2.0
babel==2.16.0
binaryornot==0.4.4
black==24.10.0
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
colorlog==6.9.0
cookiecutter==2.6.0
Django==5.1.3
django-cors-headers==4.6.0
django-unfold==0.42.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
frozenlist==1.5.0
geographiclib==2.0
geopy==2.4.1
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
Jinja2==3.1.4
jst-aicommit==0.1.6
jst-django==1.2.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.1.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
polib==1.2.0
prompt-toolkit==3.0.36
propcache==0.2.0
Pygments==2.18.0
PyJWT==2.10.1
python-click==0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2024.2
PyYAML==6.0.2
questionary==2.0.1
requests==2.32.3
rich==13.9.3
shellingham==1.5.4
six==1.16.0
sqlparse==0.5.2
tenacity==9.0.0
text-unidecode==1.3
tqdm==4.67.0
typer==0.12.5
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
wcwidth==0.2.13
whitenoise==6.3.0
yarl==1.17.1
zipfile36==0.1.3
